Make Properties.alerting a pluggable interface
Extracts Alerting as an interface with the existing concrete class renamed
to DefaultAlerting. JsonHelper.configureMapper registers DefaultAlerting
as the default abstract-type mapping for Alerting via
SimpleModule.addAbstractTypeMapping, so existing serialized snapshots
deserialize unchanged with no JSON-level differences. Downstream
consumers can register their own mapping (later-registered wins) to swap
in a different Alerting POJO without inheriting Netflix's shape.

The engine's only contract with the alerting field is update(paramParser)
for parameter interpolation; that method moves to the Alerting interface
and is overridden by DefaultAlerting. AlertingTypeConfig (one nested
SlackConfig ref) and tests are updated to use DefaultAlerting where
concrete-field access is needed.

Also extracts a small NotebookParamsContributor SPI so consumers with a
non-default Alerting impl can plug their notebook param projection in
without forking NotebookParamsBuilder. The existing alerting -> notebook-
param logic moves to DefaultAlertingNotebookParamsContributor, registered
as a Spring bean. NotebookParamsBuilder now accepts a
List<NotebookParamsContributor> and iterates.

No engine behavior changes. All existing tests pass. New AlertingTest
covers default-mapper deserialization, no extra type field on
serialization, and consumer abstract-type-mapping override round-trip.
