+++ maestro-common/src/main/java/com/netflix/maestro/models/definition/alerting/AlertingTypeConfig.java
@@ -22,6 +22,7 @@ import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.netflix.maestro.models.definition.Alerting;
+import com.netflix.maestro.models.definition.DefaultAlerting;
 import java.io.Serializable;
 import java.util.EnumSet;
 import java.util.HashSet;
@@ -42,7 +43,7 @@ public class AlertingTypeConfig implements Serializable {
   private Set<String> pagerduties;
 
   @JsonProperty("slack")
-  private Alerting.SlackConfig slackConfig;
+  private DefaultAlerting.SlackConfig slackConfig;
 
   @JsonIgnore private Set<Action> actions;
   private boolean disabled;
