+++ maestro-kubernetes/src/main/java/com/netflix/maestro/engine/notebook/NotebookParamsBuilder.java
@@ -18,14 +18,13 @@ import com.netflix.maestro.engine.execution.StepRuntimeSummary;
 import com.netflix.maestro.engine.execution.WorkflowSummary;
 import com.netflix.maestro.exceptions.MaestroBadRequestException;
 import com.netflix.maestro.models.Constants;
-import com.netflix.maestro.models.definition.Alerting;
 import com.netflix.maestro.models.definition.Step;
-import com.netflix.maestro.models.definition.Tct;
 import com.netflix.maestro.models.parameter.ParamType;
 import com.netflix.maestro.models.parameter.Parameter;
 import com.netflix.maestro.utils.Checks;
 import com.netflix.maestro.utils.MapHelper;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
@@ -51,6 +50,7 @@ public class NotebookParamsBuilder {
           Constants.INITIATOR_TIMEZONE_PARAM);
 
   private final ObjectMapper objectMapper;
+  private final List<NotebookParamsContributor> contributors;
 
   /**
    * Build notebook parameters and return the full JSON string.
@@ -73,82 +73,14 @@ public class NotebookParamsBuilder {
           workflowSummary.getCriticality().name().toLowerCase(Locale.US));
     }
     paramMap.put(NotebookConstants.ATTEMPT_NUMBER, runtimeSummary.getStepAttemptId() - 1);
-    addAlertingParamsIfPresent(paramMap, workflowSummary);
+    for (NotebookParamsContributor contributor : contributors) {
+      contributor.contribute(workflowSummary, paramMap);
+    }
 
     Map<String, Object> extraPapermillParams = generatePapermillParams(workflowSummary);
     paramMap.putAll(extraPapermillParams);
 
     return toPapermillParams(workflowSummary, paramMap);
-  }
-
-  private void addAlertingParamsIfPresent(
-      Map<String, Object> paramMap, WorkflowSummary workflowSummary) {
-    Alerting alerting = workflowSummary.getRunProperties().getAlerting();
-    if (alerting == null) {
-      return;
-    }
-    addTctParamsIfPresent(paramMap, alerting);
-    Set<String> pagerDuties = alerting.getPagerduties();
-    Set<String> emails = alerting.getEmails();
-    if (pagerDuties != null && !pagerDuties.isEmpty()) {
-      paramMap.put(NotebookConstants.PAGER_DUTIES, pagerDuties);
-    }
-    if (emails != null && !emails.isEmpty()) {
-      paramMap.put(NotebookConstants.NOTIFICATION_EMAILS, emails);
-    }
-    Alerting.SlackConfig slackConfig = alerting.getSlackConfig();
-    if (slackConfig != null) {
-      addSlackParamsIfPresent(workflowSummary, paramMap, slackConfig);
-    }
-  }
-
-  private void addSlackParamsIfPresent(
-      WorkflowSummary workflowSummary,
-      Map<String, Object> paramMap,
-      Alerting.SlackConfig slackConfig) {
-    Map<String, String> slackParams = new LinkedHashMap<>();
-    Set<String> users = slackConfig.getUsers();
-    try {
-      if (users != null && !users.isEmpty()) {
-        slackParams.put(NotebookConstants.SLACK_USERS, objectMapper.writeValueAsString(users));
-      }
-      Set<String> channels = slackConfig.getChannels();
-      if (channels != null && !channels.isEmpty()) {
-        slackParams.put(
-            NotebookConstants.SLACK_CHANNELS, objectMapper.writeValueAsString(channels));
-      }
-    } catch (JsonProcessingException e) {
-      String workflowIdentity = workflowSummary.getIdentity();
-      LOG.warn("Failed to serialize slack params for workflow {}", workflowIdentity, e);
-      throw new MaestroBadRequestException(
-          e, "Failed to serialize slack params for workflow %s", workflowIdentity);
-    }
-    if (!slackParams.isEmpty()) {
-      paramMap.put(NotebookConstants.SLACK_PARAM, slackParams);
-    }
-  }
-
-  private void addTctParamsIfPresent(Map<String, Object> paramMap, Alerting alerting) {
-    Tct tct = alerting.getTct();
-    if (tct == null) {
-      return;
-    }
-    Map<String, String> tctMap = new LinkedHashMap<>();
-    if (tct.getCompletedByHour() != null) {
-      tctMap.put(NotebookConstants.TCT_COMPLETED_BY_HOUR, tct.getCompletedByHour().toString());
-    }
-    if (tct.getDurationMinutes() != null) {
-      tctMap.put(NotebookConstants.TCT_DURATION_MINUTES, tct.getDurationMinutes().toString());
-    }
-    if (tct.getCompletedByTs() != null) {
-      tctMap.put(NotebookConstants.TCT_COMPLETED_BY_TS, tct.getCompletedByTs().toString());
-    }
-    if (tct.getTz() != null) {
-      tctMap.put(NotebookConstants.TCT_TZ, tct.getTz());
-    }
-    if (!tctMap.isEmpty()) {
-      paramMap.put(NotebookConstants.TCT_PARAM, tctMap);
-    }
   }
 
   /**
